Reframe test enhancement #51 - local config deleted - set up for gaya
